# openapi/paths.yaml
openapi: 3.0.3
info:
  title: Swagger Petstore - OpenAPI 3.0
  description: |-
    This is a sample Pet Store Server based on the OpenAPI 3.0 specification.  You can find out more about
    Swagger at [https://swagger.io](https://swagger.io). In the third iteration of the pet store, we've switched to the design first approach!
    You can now help us improve the API whether it's by making changes to the definition itself or to the code.
    That way, with time, we can improve the API in general, and expose some of the new features in OAS3.

    _If you're looking for the Swagger 2.0/OAS 2.0 version of Petstore, then click [here](https://editor.swagger.io/?url=https://petstore.swagger.io/v2/swagger.yaml). Alternatively, you can load via the `Edit > Load Petstore OAS 2.0` menu option!_

    Some useful links:
    - [The Pet Store repository](https://github.com/swagger-api/swagger-petstore)
    - [The source API definition for the Pet Store](https://github.com/swagger-api/swagger-petstore/blob/master/src/main/resources/openapi.yaml)
  termsOfService: http://swagger.io/terms/
  contact:
    email: apiteam@swagger.io
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
...
servers:
  - url: http://localhost:4000
...
paths:
...
  /users/me:
    get:
      tags:
        - users
      summary: Lấy thông tin user
      description: Lấy thông tin user
      operationId: getMe
      security:
        - BearerAuth: []
      responses:
        default:
          description: Lấy thông tin user thành công
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
    patch:
      tags:
        - users
      summary: Cập nhật thông tin user
      description: Cập nhật thông tin user
      operationId: updateMe
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin cập nhật user
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateMeBody'
        required: true
      responses:
        default:
          description: Cập nhật thông tin user thành công
...
  /users/refresh-token:
    post:
      tags:
        - users
      summary: Lấy lại access token
      description: Lấy lại access token
      operationId: refreshToken
      requestBody:
        description: Thông tin lấy lại access token
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
        required: true
      responses:
        '200':
          description: Lấy lại access token thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Refresh token success
                  result:
                    $ref: '#/components/schemas/SuccessAuthentication'
  /users/resend-verify-email:
    post:
      tags:
        - users
      summary: Gửi lại email xác thực
      description: Gửi lại email xác thực
      operationId: resendVerifyEmail
      requestBody:
        description: Thông tin gửi lại email xác thực
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Gửi lại email xác thực thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Resend verify email success
  /users/forgot-password:
    post:
      tags:
        - users
      summary: Quên mật khẩu
      description: Quên mật khẩu
      operationId: forgotPassword
      requestBody:
        description: Thông tin quên mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  example: duthanhduoc100@gmail.com
        required: true
      responses:
        '200':
          description: Nhận được email reset password
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Kiểm tra email để reset mật khẩu
...
  /users/reset-password:
    post:
      tags:
        - users
      summary: Reset mật khẩu
      description: Reset mật khẩu
      operationId: resetPassword
      requestBody:
        description: Thông tin reset mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                forgot_password_token:
                  type: string
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjRiZTBhZDJlNDNkMjQ2NDM5NGZlZWRiIiwidG9rZW5fdHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTEzODMyMjYsImV4cCI6MTcwMDAyMzIyNn0.bFnaBfxWl-Q48HCwAeMpNzZwTO9LEjcBd7Zyipjqr64
                password:
                  type: string
                  example: Duoc123!
                confirm_password:
                  type: string
                  example: Duoc123!
        required: true
      responses:
        default:
          description: Reset mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Reset mật khẩu thành công
...
  /users/change-password:
    put:
      tags:
        - users
      summary: Đổi mật khẩu
      description: Đổi mật khẩu
      operationId: changePassword
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin đổi mật khẩu
        content:
          application/json:
            schema:
              type: object
              properties:
                old_password:
                  type: string
                  example: Duoc123!
                password:
                  type: string
                  example: Duoc1234!
                confirm_password:
                  type: string
                  example: Duoc1234!
        required: true
      responses:
        default:
          description: Đổi mật khẩu thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Đổi mật khẩu thành công
  /tweets:
    post:
      tags:
        - tweets
      summary: Tạo tweet
      description: Tạo tweet
      operationId: createTweet
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin tạo tweet
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TweetRequestBody'
        required: true
      responses:
        default:
          description: Tạo tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Tạo tweet thành công
    get:
      tags:
        - tweets
      summary: Lấy danh sách tweet từ những người mình theo dõi
      description: Lấy danh sách tweet từ những người mình theo dõi
      operationId: getTweets
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          description: Số trang
          required: false
          schema:
            type: integer
            example: 1
        - name: limit
          in: query
          description: Số lượng tweet trên 1 trang
          required: false
          schema:
            type: integer
            example: 10
      responses:
        default:
          description: Lấy danh sách tweet thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Lấy danh sách tweet thành công
                  result:
                    type: object
                    properties:
                      tweets:
                        type: array
                        items:
                          $ref: '#/components/schemas/Tweet'
                      total_page:
                        type: integer
                        example: 1
                      limit:
                        type: integer
                        example: 10
                      page:
                        type: integer
                        example: 1
...
  /medias/upload-image:
    post:
      tags:
        - medias
      summary: Upload ảnh
      description: Upload ảnh
      operationId: uploadImage
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload ảnh, upload 1 hoặc nhiều ảnh
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                image:
                  type: array
                  items:
                    type: string
                    format: binary
                  maxItems: 4
        required: true
      responses:
        default:
          description: Upload ảnh thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload ảnh thành công
                  result:
                    type: array
                    items:
                      $ref: '#/components/schemas/Media'
  /medias/upload-video:
    post:
      tags:
        - medias
      summary: Upload video
      description: Upload video
      operationId: uploadVideo
      security:
        - BearerAuth: []
      requestBody:
        description: Thông tin upload video, chỉ upload được 1 video
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
        required: true
      responses:
        default:
          description: Upload video thành công
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Upload video thành công
                  result:
                    $ref: '#/components/schemas/Media'
...
components:
  schemas:
...
    SuccessAuthentication:
      type: object
      properties:
        access_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwY2ZiZjhkOWM2YjE5ODk2OWRkMDM3IiwidHlwZSI6MCwidmVyaWZ5IjoxLCJpYXQiOjE2OTU4NjY1OTUsImV4cCI6MTY5NTk1Mjk5NX0.Ms9hMxU80GSgvSlR0J1c5SFTP05IAn8hFxbA3r4fV8o
        refresh_token:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoiNjUwY2ZiZjhkOWM2YjE5ODk2OWRkMDM3IiwidHlwZSI6MSwidmVyaWZ5IjoxLCJpYXQiOjE2OTU4NjY1OTUsImV4cCI6MTcwNDUwNjU5NX0.SKcyZ3Wjd5u5CiKbJgbDl5oU__X4_JS9RfVfUvmFHw8
    User:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          example: '650cfbf8d9c6b198969dd037'
        name:
          type: string
          example: 'Nam'
        email:
          type: string
          example: 'namskymtp23+9@gmail.com'
        date_of_birth:
          type: string
          format: ISO8601
          example: 2023-06-30T09:15:13.355Z
        created_at:
          type: string
          format: ISO8601
          example: 2023-09-22T02:29:12.552Z
        updated_at:
          type: string
          format: ISO8601
          example: 2023-09-22T02:29:12.552Z
        verify:
          $ref: '#/components/schemas/UserVerifyStatus'
        twitter_circle:
          type: array
          items:
            type: string
            format: MongoId
          example: ['650cfbf8d9c6b198969dd037', '650cfbf8d9c6b198969dd038']
        bio:
          type: string
          example: 'This is my bio'
        location:
          type: string
          example: 'San Francisco, CA'
        website:
          type: string
          example: 'www.namskymtp23.com'
        username:
          type: string
          example: 'user650cfbf8d9c6b198969dd037'
        avatar:
          type: string
          example: 'http://localhost:4000/images/avatars/namvan.jpg'
        cover_photo:
          type: string
          example: 'http://localhost:4000/images/avatars/namvan.jpg'
    UserVerifyStatus:
      type: number
      enum: [Unverified, Verified, Banned]
      example: 1
...
    UpdateMeBody:
      type: object
      properties:
        name:
          type: string
          example: John Doe
        date_of_birth:
          type: string
          format: ISO8601
          example: 2023-06-08T10:17:31.096Z
        bio:
          type: string
          example: 'This is my bio.'
        location:
          type: string
          example: 'San Francisco, CA'
        website:
          type: string
          example: 'www.example.com'
        username:
          type: string
          example: 'johndoe'
        avatar:
          type: string
          example: 'http:localhost:4000/images/avatars/johndoe.jpg'
        cover_photo:
          type: string
          example: 'http:localhost:4000/images/avatars/johndoe.jpg'
    TweetRequestBody:
      type: object
      properties:
        type:
          $ref: '#/components/schemas/TweetType'
        audience:
          $ref: '#/components/schemas/TweetAudience'
        content:
          type: string
          description: Nội dung tweet
          example: 'This is my tweet.'
        parent_id:
          type: string
          nullable: true
          description: ID của tweet cha
          example: 64be0ad2e43d2464394feedb
        hashtags:
          type: array
          items:
            type: string
          description: Danh sách hashtag
          example: ['hashtag1', 'hashtag2']
        mentions:
          type: array
          items:
            type: string
          description: Danh sách user_id được đề cập
          example: ['64be0ad2e43d2464394feedb', '64be0ad2e43d2464394feedc']
        medias:
          type: array
          items:
            $ref:
          description: Danh sách media đính kèm
    TweetType:
      type: number
      enum: [Tweet, Retweet, Comment, QuoteTweet]
      example: 1
    TweetAudience:
      type: number
      enum: [Everyone, TwitterCircle]
      example: 1
...
    MediaType:
      type: number
      enum: [Image, Video, HLS]
      example: 1
    Media:
      type: object
      properties:
        url:
          type: string
          example: 'http://localhost:4000/images/avatars/johndoe.jpg'
        type:
          $ref: '#/components/schemas/MediaType'
    Tweet:
      type: object
      properties:
        _id:
          type: string
          format: MongoId
          description: ID của tweet
          example: 64be0ad2e43d2464394feedb
        user_id:
          type: string
          format: MongoId
          description: ID của người dùng tạo tweet
          example: 64be0ad2e43d2464394feedb
        type:
          $ref: '#/components/schemas/TweetType'
        audience:
          $ref: '#/components/schemas/TweetAudience'
        content:
          type: string
          description: Nội dung tweet
          example: 'This is my tweet.'
        parent_id:
          type: string
          nullable: true
          format: MongoId
          description: ID của tweet cha
          example: 64be0ad2e43d2464394feedb
        hashtags:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                format: MongoId
                description: ID của hashtag
                example: 64be0ad2e43d2464394feedb
              name:
                type: string
                description: Tên của hashtag
                example: 'hashtag1'
              created_at:
                type: string
                format: ISO8601
                description: Thời gian tạo hashtag
                example: 2023-03-08T12:00:00Z
          description: Danh sách hashtag
          example:
            [
              { _id: '64be0ad2e43d2464394feedb', name: 'hashtag1', created_at: '2023-03-08T12:00:00Z' },
              { _id: '64be0ad2e43d2464394feedc', name: 'hashtag2', created_at: '2023-03-08T12:00:00Z' }
            ]
        mentions:
          type: array
          items:
            type: object
            properties:
              _id:
                type: string
                format: MongoId
                description: ID của user được đề cập
                example: 64be0ad2e43d2464394feedb
              name:
                type: string
                description: Tên của user được đề cập
                example: 'John Doe'
              username:
                type: string
                description: Username của user được đề cập
                example: 'johndoe'
              email:
                type: string
                description: Email của user được đề cập
                example: duthanhduoc@gmail.com
        medias:
          type: array
          items:
            $ref: '#/components/schemas/Media'
          description: Danh sách media đính kèm
          example:
            [
              { url: 'http://localhost:4000/images/avatars/johndoe.jpg', type: 1 },
              { url: 'http://localhost:4000/images/avatars/johndoe.jpg', type: 2 }
            ]
        retweet_count:
          type: integer
          description: Số lần retweet
        comment_count:
          type: integer
          description: Số lần comment
        quote_count:
          type: integer
          description: Số lần quote
        bookmarks:
          type: integer
          description: Số lần bookmark
        likes:
          type: integer
          description: Số lần like
        guest_views:
          type: integer
          description: Số lần xem của khách
        user_views:
          type: integer
          description: Số lần xem của người dùng
        created_at:
          type: string
          format: date-time
          description: Thời gian tạo tweet
        updated_at:
          type: string
          format: date-time
          description: Thời gian cập nhật tweet
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT